Cleaned up file structure & added heroku runtime.txt
Moved most of the source files to src/, more organization
to be done at some point but just need to tidy it up for
now. Also tweaked start.sh to reflect this change.

Added runtime.txt to point to Python 3 as Heroku doesn't
default to it, breaking end="" and therefore crashing the
app.

Added servers.json with official GGServers data for deployment
and also so you can test it at home.