Python panel-side script complete, working on PHP
The python script for deployment on panels is complete.
All that it really requires is Python 3.x to be installed
and screen for whatever distro the panel is running. Also
supports Windows but the Run.sh script obviously does
not. Work started on PHP script, trying to merge as much
as possible to avoid seperate Windows / Linux versions but
might be unavoidable.