json/ -> cache/, sleeker index.html/.css
Renamed the json directory to cache to accomadate for the new cache
module, and also changed the vars passed to Fetch and Stash functions
to match that in the cache module. Transitioned index page to use a
table instead of a list, which makes it look much more attractive.
Also small tweak to make it more obvious when a server is offline.