Basic config and converting a lot to OOP standards.
Implemented a basic config file for various options and also API keys.
Also converted a larger portion of the codebase to use OOP stadards
(classes, etc). Implemented App.py which is the new brain of the app
and interprets the config and decide what frontends to launch. Scanning
is currently broken however I am investigating a fix - it seems flask is
blocking the execution despite being a different process thread. *shrug*.