Working on websocket implementation for fetch
Decided to port /fetch to websockets (legacy API path will be added later).
However currently have strange bug in platypus.js in which the wrong row will
be modified - seems to be whichever row is set in RequestStat() instead of the
row passed through to the callback. Run 'python src/Tornado.py' to start the
Tornado webserver.