Starting work on new Active Online Reporting
Implemented the barebones ServerHandler for AOR and to replace SQL.py.
Aiming to move away from a pure SQL solution to a mix of a json cache
for stats and SQL for server metadata (uuid, ip, hostname, name, etc).
Also working on new admin interface to make it easier to manage a larger
selection of servers.