Fixed some SQL queries and fully implemented AOR
AOR is now completely implemented. Sort of. The client
script isn't quite done but I've written all the backend
code for the master to provide a socket and aggregate the
data. data.json has some silly placeholder stuff, I recommend
running Scripts/python/convert-to-serverhandler.py (modify
it before so it can connect to your db and whatnot). Working
on implementing pushing of data to client frontend and a better
conversion script. Also an updated setup.py.